+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 ## 0.1.0-dev1
 
+* Added a /healthcheck endpoint
 * Added an API endpoint for detectron2 layout detection